MISC Refractor Class WokerFtTask
