+++ src/simulationPushTask.ts
@@ -1,5 +1,5 @@
 import {CloudConfig, CloudProvider, PublishFileData} from 'packunpackservice';
-import {UploadTask} from './helpers/workerFtTask';
+import {UploadTask} from './helpers/Tasks/UploadTask';
 import CloudManager from './services/cloudManager.services';
 import {Types} from 'mongoose';
 let numtaskIndex = 0; // this help monitor the nums of task was sent
